Adds support for Bugello (Bugzilla -> Trello)
This is totally procedural and fairly tailored for the Wikimedia
Foundation's mobile apps team. Using the same general principles as
Mingle, this will take bugs retrieved from a Bugzilla feed and create
Trello cards in a defined Trello list on a matched Trello board.

Right now, this expects Trello boards to be created per-iteration,
following a pattern of "Some generic Title - Sprint <some #>" and will
automatically pick the Trello board with the highest trailing #.

All of this should be refactored, cleaned up, and generecized, but this
is usable now!