Use 'last_change_time' rather than 'creation_time' in query for bugs
Addressses issue #17.

This may result in 'false positivies' (in as much as we're looking for
'new' bugs and this will return 'old' bugs that have been updated), but
has the benefit of picking up bugs that have been moved from a component
that you may not query for into a component that you do, but was created
prior to the last time bingle ran.

Further, this could be useful for auto-update existing bug cards in
Mingle with new information from Bugzilla.