Adds bingleResolved.py to auto-handle bugs resolved in BZ with Mingle
This is a workable proof of concept for taking actions on bug cards in
Mingle when a bug in Bugzilla is marked as resolved. Currently, card
statuses are hardcoded which makes this not super useful unless your
Mingle bug card statuses are the same as the WMF Mobile Web team's - but
it shouldn't take much to abstract this out and make it configruable.

Ultimately, it would be nice to make this generic - such that you can
define a mapping of certain bugzilla statuses to define actions to
automatically take on cards in Mingle (eg putting reopened bugs into
needs analysis, or something).